backend/requirements.txt
fastapi==0.114.0
httpx==0.27.2
uvicorn==0.30.6
pydantic-settings==2.4.0
